§ &§&•–_THOMASjosephÇatalano—§&&§•&§ &§& .tomcatalano.0@gmail.com .tomcatalano.0@gmail.com&§& ## Please do not write below this line ##  Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.  Ticket ID: 1022505   If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)  Thank you! The GitHub Support Team  This email is a service from GitHub Support. [DOYQLZ-QQ90]§&§©️2021&§&.§™️ ## Please do not write below this line ##  Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.  Ticket ID: 1022505   If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)  Thank you! The GitHub Support Team  This email is a service from GitHub Support. [DOYQLZ-QQ90]&§&.§&§–THOMASjosephÇatalano—•&§&,.&§& ## Please do not write below this line ##  Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.  Ticket ID: 1022505   If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)  Thank you! The GitHub Support Team  This email is a service from GitHub Support. [DOYQLZ-QQ90]§&§©️®️™️§&§.&—…–   ||||_•”’THOMAS-j._Catalano’*”’•_||||§&
&§ &§& .tomcatalano.0@gmail.com .tomcatalano.0@gmail.com&§& ## Please do not write below this line 
Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.

Ticket ID: 1022505


If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)

Thank you!
The GitHub Support Team

This email is a service from GitHub Support.
[DOYQLZ-QQ90]§&§©️2021&§&.§™️ ## Please do not write below this line ##

Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.

Ticket ID: 1022505


If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)

Thank you!
The GitHub Support Team

This email is a service from GitHub Support.
[DOYQLZ-QQ90]&§&.§&§–THOMASjosephÇatalano—•&§&,.&§& ## Please do not write below this line ##

Thank you for contacting GitHub Support. We wanted to let you know that we've received your message. In order to respond to tickets with the greatest urgency as quickly as possible during the COVID-19 crisis, we've established a priority order. If you have questions regarding our recent announcement that makes most GitHub features accessible to our community free of charge, we have captured answers to common questions here.

Ticket ID: 1022505


If you have any additional information or would like to add anything to your initial message, now would be a great time to do so, feel free to reply to this email. If not, then rest assured your request is in the right hands :)

Thank you!
The GitHub Support Team

This email is a service from GitHub Support.
[DOYQLZ-QQ90]§&§©️®️™️§&§.&—…–


||||_•”’THOMAS-j._Catalano’*”’•_||||§& &§&•–_THOMASjosephÇatalano—§&&§•